@@ -8,16 +8,25 @@ def keystoreProperties = new Properties()
 keystoreProperties.load(new FileInputStream(keystorePropertiesFile))
 
 android {
-  compileSdkVersion 27
+  compileSdkVersion 28
   defaultConfig {
     applicationId "com.raywenderlich.android.datadrop"
     minSdkVersion 19
-    targetSdkVersion 27
+    targetSdkVersion 28
     versionCode 1
     versionName "1.0"
-    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
 
     resValue 'string', 'GOOGLE_MAPS_API_KEY', keystoreProperties['GOOGLE_MAPS_API_KEY']
+
+    javaCompileOptions {
+      annotationProcessorOptions {
+        arguments += [
+                "room.schemaLocation":"2.2.5",
+                "room.incremental":"true",
+                "room.expandProjection":"true"]
+      }
+    }
   }
   buildTypes {
     release {
@@ -29,23 +38,31 @@ android {
 
 dependencies {
   implementation fileTree(include: ['*.jar'], dir: 'libs')
-  implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
-  implementation "com.android.support:appcompat-v7:$support_lib_version"
-  implementation "com.android.support:design:$support_lib_version"
-  implementation 'com.google.android.gms:play-services-maps:11.8.0'
-  implementation 'com.android.support.constraint:constraint-layout:1.0.2'
+  implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+  //noinspection GradleDependency
+  implementation 'androidx.appcompat:appcompat:1.0.0'
+  //noinspection GradleDependency
+  implementation 'com.google.android.material:material:1.0.0'
+  implementation 'com.google.android.gms:play-services-maps:17.0.0'
+  implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
   testImplementation 'junit:junit:4.12'
-  androidTestImplementation 'com.android.support.test:runner:1.0.1'
-  androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
-  implementation 'com.google.code.gson:gson:2.8.2'
+  androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+  androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+  implementation 'com.google.code.gson:gson:2.8.6'
 
-  implementation "android.arch.persistence.room:runtime:$room_version"
-  annotationProcessor "android.arch.persistence.room:compiler:$room_version"
-  kapt "android.arch.persistence.room:compiler:$room_version"
-  androidTestImplementation "android.arch.persistence.room:testing:$room_version"
+  //noinspection GradleDependency
+  implementation 'androidx.room:room-runtime:2.0.0'
+  annotationProcessor 'androidx.room:room-compiler:2.2.5'
+  kapt 'androidx.room:room-compiler:2.2.5'
+  //noinspection GradleDependency
+  androidTestImplementation 'androidx.room:room-testing:2.0.0'
+  implementation "androidx.room:room-ktx:2.2.5"
 
-  implementation "android.arch.lifecycle:extensions:$archLifecycle_version"
-  annotationProcessor "android.arch.lifecycle:compiler:$archLifecycle_version"
+  //noinspection GradleDependency,GradleDependency
+  implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+  annotationProcessor 'androidx.lifecycle:lifecycle-compiler:2.2.0'
 
+  implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+  implementation "androidx.cardview:cardview:1.0.0"
 
 }
