@@ -2,8 +2,8 @@
 
 buildscript {
   ext {
-    kotlin_version = '1.2.21'
-    support_lib_version = '27.0.2'
+    kotlin_version = '1.3.72'
+    support_lib_version = '28.0.0'
     room_version = '1.0.0'
     archLifecycle_version = '1.1.0'
   }
@@ -12,7 +12,7 @@ buildscript {
     jcenter()
   }
   dependencies {
-    classpath 'com.android.tools.build:gradle:3.0.1'
+    classpath 'com.android.tools.build:gradle:4.0.0'
     classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
     // NOTE: Do not place your application dependencies here; they belong
