@@ -1,8 +1,8 @@
 package com.raywenderlich.android.datadrop.model
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.RoomDatabase
-import android.arch.persistence.room.TypeConverters
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import androidx.room.TypeConverters
 
 
 @Database(entities = [(Drop::class) , (MarkerColor::class)], version = 1)
